Update relevant unit tests
